+++ Cargo.toml
@@ -18,6 +18,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 [workspace.dependencies]
 alloy-consensus = { version = "0.1.0", path = "crates/consensus" }
+alloy-eips = { version = "0.1.0", path = "crates/eips" }
 alloy-json-rpc = { version = "0.1.0", path = "crates/json-rpc" }
 alloy-network = { version = "0.1.0", path = "crates/network" }
 alloy-pubsub = { version = "0.1.0", path = "crates/pubsub" }
@@ -37,6 +38,7 @@ alloy-sol-types = { version = "0.5.1", default-features = false, features = ["st
 alloy-rlp = "0.3"
 
 # crypto
+c-kzg = "0.4.0" # for eip-4844
 elliptic-curve = { version = "0.13.5", default-features = false, features = ["std"] }
 generic-array = { version = "0.14.7", default-features = false, features = ["std"] }
 k256 = { version = "0.13.2", default-features = false, features = ["ecdsa", "std"] }
@@ -59,12 +61,14 @@ tracing-subscriber = "0.3.18"
 
 tempfile = "3.8"
 
+# misc
 assert_matches = "1.5"
 auto_impl = "1.1"
 base64 = "0.21"
 bimap = "0.6"
 home = "0.5"
 itertools = "0.12"
+once_cell = "1.17"
 pin-project = "1.1"
 rand = "0.8.5"
 reqwest = { version = "0.11.18", default-features = false }
@@ -75,6 +79,8 @@ serde_with = "3.4"
 serial_test = "2.0"
 thiserror = "1.0"
 url = "2.4"
+
+
 
 # testing
 arbitrary = "1.1"
