+++ crates/consensus/Cargo.toml
@@ -16,6 +16,7 @@ alloy-rlp.workspace = true
 serde = { workspace = true, features = ["derive"], optional = true }
 
 arbitrary = { workspace = true, features = ["derive"], optional = true }
+alloy-eips.workspace = true
 
 [dev-dependencies]
 hex-literal = "0.4.1"
