+++ crates/consensus/src/header.rs
@@ -7,10 +7,9 @@ use alloy_rlp::{
     EMPTY_STRING_CODE,
 };
 
-use crate::{
-    calc_next_block_base_fee,
-    pure::{calc_blob_gasprice, calc_excess_blob_gas},
-    BaseFeeParams,
+use alloy_eips::{
+    eip1559::{calc_next_block_base_fee, BaseFeeParams},
+    eip4844::{calc_blob_gasprice, calc_excess_blob_gas},
 };
 
 /// Ommer root of empty list.
