+++ crates/eips/Cargo.toml
@@ -1,7 +1,5 @@
 [package]
-name = "alloy-network"
-description = "Ethereum blockchain RPC behavior abstraction"
-
+name = "alloy-eips"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -11,12 +9,14 @@ homepage.workspace = true
 repository.workspace = true
 exclude.workspace = true
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [dependencies]
-alloy-eips = { workspace = true, features = ["serde"] }
-alloy-json-rpc.workspace = true
 alloy-primitives.workspace = true
 alloy-rlp.workspace = true
-serde = { workspace = true, features = ["derive"] }
+
+serde = { workspace = true, optional = true }
+thiserror.workspace = true
 
 [features]
-k256 = ["alloy-primitives/k256"]
+serde = ["dep:serde", "alloy-primitives/serde"]
