+++ crates/eips/src/eip4788.rs
@@ -1,3 +1,7 @@
+//! [EIP-4788] constants.
+//!
+//! [EIP-4788]: https://eips.ethereum.org/EIPS/eip-4788
+
 use alloy_primitives::{address, Address};
 
 /// The caller to be used when calling the EIP-4788 beacon roots contract at the beginning of the
