+++ crates/eips/src/eip4844.rs
@@ -1,6 +1,36 @@
-use crate::constants::eip4844::{
-    BLOB_GASPRICE_UPDATE_FRACTION, BLOB_TX_MIN_BLOB_GASPRICE, TARGET_DATA_GAS_PER_BLOCK,
-};
+//! [EIP-4844] constants and helpers.
+//!
+//! [EIP-4844]: https://eips.ethereum.org/EIPS/eip-4844
+
+/// Size a single field element in bytes.
+pub const FIELD_ELEMENT_BYTES: u64 = 32;
+
+/// How many field elements are stored in a single data blob.
+pub const FIELD_ELEMENTS_PER_BLOB: u64 = 4096;
+
+/// Gas consumption of a single data blob.
+pub const DATA_GAS_PER_BLOB: u64 = 131_072u64; // 32*4096 = 131072 == 2^17 == 0x20000
+
+/// Maximum data gas for data blobs in a single block.
+pub const MAX_DATA_GAS_PER_BLOCK: u64 = 786_432u64; // 0xC0000 = 6 * 0x20000
+
+/// Target data gas for data blobs in a single block.
+pub const TARGET_DATA_GAS_PER_BLOCK: u64 = 393_216u64; // 0x60000 = 3 * 0x20000
+
+/// Maximum number of data blobs in a single block.
+pub const MAX_BLOBS_PER_BLOCK: usize = (MAX_DATA_GAS_PER_BLOCK / DATA_GAS_PER_BLOB) as usize; // 786432 / 131072  = 6
+
+/// Target number of data blobs in a single block.
+pub const TARGET_BLOBS_PER_BLOCK: u64 = TARGET_DATA_GAS_PER_BLOCK / DATA_GAS_PER_BLOB; // 393216 / 131072 = 3
+
+/// Determines the maximum rate of change for blob fee
+pub const BLOB_GASPRICE_UPDATE_FRACTION: u64 = 3_338_477u64; // 3338477
+
+/// Minimum gas price for a data blob
+pub const BLOB_TX_MIN_BLOB_GASPRICE: u128 = 1u128;
+
+/// Commitment version of a KZG commitment
+pub const VERSIONED_HASH_VERSION_KZG: u8 = 0x01;
 
 /// Calculates the `excess_blob_gas` from the parent header's `blob_gas_used` and `excess_blob_gas`.
 ///
@@ -57,7 +87,6 @@ fn fake_exponential(factor: u64, numerator: u64, denominator: u64) -> u128 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::constants::eip4844::DATA_GAS_PER_BLOB;
 
     // https://github.com/ethereum/go-ethereum/blob/28857080d732857030eda80c69b9ba2c8926f221/consensus/misc/eip4844/eip4844_test.go#L27
     #[test]
