+++ crates/eips/src/lib.rs
@@ -1,5 +1,13 @@
-mod eip1559;
+pub mod eip1559;
 pub use eip1559::calc_next_block_base_fee;
 
-mod eip4844;
+pub mod eip2718;
+
+pub mod eip2930;
+
+pub mod eip4788;
+
+pub mod eip4844;
 pub use eip4844::{calc_blob_gasprice, calc_excess_blob_gas};
+
+pub mod merge;
